fix : 매우 큰 정수 계산이 가능하게 변경

+++ src/main/java/calculator/io/ConsoleOutputHandler.java
@@ -1,12 +1,14 @@
 package calculator.io;
 
+import java.math.BigInteger;
+
 public class ConsoleOutputHandler {
 
     public void showInputComments(){
         System.out.println("덧셈할 문자열을 입력해 주세요.");
     }
 
-    public void showFinalOutput(int sum) {
+    public void showFinalOutput(BigInteger sum) {
         System.out.println("결과 : " + sum);
     }
 }
